requirements.txt
flask
flask-cors
pillow
imagehash
joblib
scikit-learn
pandas
numpy
 